requirements.txt
gunicorn==22.0.0
Flask==3.0.3
Flask-Cors==4.0.1
Werkzeug==3.0.3
google-generativeai==0.5.2
requests==2.31.0